% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generic.R
\name{as.data.frame.Dataset}
\alias{as.data.frame.Dataset}
\title{Convert Dataset to data.frame}
\usage{
\method{as.data.frame}{Dataset}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{Dataset to convert}

\item{row.names}{Unused for now, needed for consistency with generic}

\item{optional}{Unused for now, needed for consistency with generic}

\item{\dots}{Unused for now, needed for consistency with generic}
}
\value{
A data frame with the same contents as the dataset
}
\description{
Convert Dataset to data.frame
}
\examples{
\dontrun{
ws = workspace(id, authorization.token)
as.data.frame(ws$datasets[[1]])
}
}
\seealso{
Other dataset functions: \code{\link{[[.Datasets}};
  \code{\link{as.list.Datasets}}; \code{\link{datasets}};
  \code{\link{names.Datasets}}
}

