% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generic.R
\name{as.list.Datasets}
\alias{as.list.Datasets}
\title{Convert Datsets object to list}
\usage{
\method{as.list}{Datasets}(x, ...)
}
\value{
A list of datasets
}
\description{
Convert Datsets object to list
}
\seealso{
Other dataset functions: \code{\link{[[.Datasets}};
  \code{\link{as.data.frame.Dataset}};
  \code{\link{datasets}}; \code{\link{names.Datasets}}
}

