% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generic.R
\name{as.list.Experiments}
\alias{as.list.Experiments}
\title{Convert Collection of Datasets and Experiments to lists}
\usage{
\method{as.list}{Experiments}(x, ...)
}
\arguments{
\item{x}{The object to convert to a list}

\item{\dots}{Unused, just for compatibility with generic}
}
\value{
A Dataset or Experiment object
}
\description{
Convert Collection of Datasets and Experiments to lists
}
\examples{
\dontrun{
ws = workspace(id, authorization.token)
as.list(experiments(ws))
as.list(datasets(ws))
}
}
\seealso{
Other experiment functions: \code{\link{dataset}};
  \code{\link{experiments}}
}

