% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/azureml-deprecated.R
\name{consumeDataframe}
\alias{consumeDataframe}
\alias{consumeFile}
\alias{consumeLists}
\title{Deprecated functions}
\usage{
consumeDataframe()

consumeFile()

consumeLists()
}
\description{
Deprecated functions
}

