% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/consume.R
\name{callAPI}
\alias{callAPI}
\title{Framework for making an Azure ML web service API call.}
\usage{
callAPI(apiKey, requestUrl, keyvalues, globalParam, retryDelay)
}
\arguments{
\item{apiKey}{primary API key}

\item{requestUrl}{API URL}

\item{keyvalues}{the data to be passed to the web service}

\item{globalParam}{the global parameters for the web service}

\item{retryDelay}{number of seconds to wait after failing (max 3 tries) to try again}
}
\value{
result the response
}
\description{
Helper function that constructs and send the API call to a Microsoft Azure Machine Learning web service, then receives and returns the response in JSON format.
}
\keyword{internal}

