% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/consume.R
\name{consumeLists}
\alias{consumeLists}
\title{Use a web service to score data in list (key=value) format}
\usage{
consumeLists(endpoint, ..., globalParam, retryDelay = 10)
}
\arguments{
\item{endpoint}{AzureML Web Service endpoint returned by \code{\link{endpoints}}}

\item{...}{variable number of requests entered as lists in key-value format}

\item{globalParam}{global parameters entered as a list, default value is an empty list}

\item{retryDelay}{the time in seconds to delay before retrying in case of a server error}
}
\value{
data frame containing results returned from web service call
}
\description{
Score data represented as lists where each list key represents
a parameter of the web service.
}
\examples{
\dontrun{
# Use a default configuration in ~/.azureml/config, alternatively
# see help for `workspace`.
ws <- workspace()

# Really simple example:
add <- function(x,y) x + y
endpoint <- publishWebService(ws, add, "add_service",
              list(x="numeric", y="numeric"), list(ans="numeric"))
result <- consumeLists(endpoint, list(x=pi, y=2))
result$ans

# A neat trick to evaluate any expression  and view its output:
p <- publishWebService(ws, fun=function(expr) {
       paste(capture.output(eval(parse(text=expr))), collapse="\\n")},
       name="commander", inputSchema=list(x="character"),
       outputSchema=list(ans="character"))
cat(consumeList(ep, list=(expr="getwd()"))$ans)
cat(consumeList(ep, list=(expr=".packages(all=TRUE)"))$ans)

# The following example illustrates scoping rules. Note that the function
# refers to the variable y defined outside the function body. That value
# will be exported with the service.
y <- pi
ep <- publishWebService(ws, fun=function(x) x + y, name="lexi",
        inputSchema=list(x="numeric"), outputSchema=list(ans="numeric"))
cat(consumeList(ep, list(x=2))$ans)
}
}
\seealso{
\code{\link{publishWebService}} \code{\link{endpoints}} \code{\link{services}} \code{\link{workspace}}

Other consumption functions: \code{\link{workspace}}
}

