% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runconfig.R
\name{create_script_run_config}
\alias{create_script_run_config}
\title{Create script runconfig}
\usage{
create_script_run_config(source_directory, script = NULL,
  arguments = NULL, target = NULL, data_references = NULL,
  cpu_image = "himanshuaml/aml-r")
}
\arguments{
\item{source_directory}{source directory containing the script}

\item{script}{script name}

\item{arguments}{arguments to script}

\item{target}{compute target}

\item{data_references}{list of data references}

\item{cpu_image}{image to use}
}
\value{
script runconfig object
}
\description{
Create script runconfig
}
