\name{dataset}
\alias{dataset}
\title{
Extract datasets from experiments}
\description{
Extract datasets from experiments}
\usage{
dataset(x, node, port, data.type)
}
\arguments{
  \item{x}{Experiment from which to extract a dataset}
  \item{node}{Node id}
  \item{port}{Port name}
  \item{data.type}{Format of intermediate data set}
}
\value{
A remote dataset}
