% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/delete_dataset.R
\name{delete.datasets}
\alias{delete.datasets}
\title{Delete datasets from an AzureML workspace.}
\usage{
delete.datasets(ws, name)
}
\arguments{
\item{name}{Either one or more \code{Dataset} objects (rows from the workspace
\code{datasets} data.frame), or a character vector of dataset names to delete.}
}
\value{
A data.frame with columns Name, Deleted, status_code indicating the HTTP
status code and success/failure result of the delete operation for each dataset.
}
\description{
Delete datasets from an AzureML workspace.
}
\seealso{
Other dataset functions: \code{\link{datasets}};
  \code{\link{download.intermediate.dataset}};
  \code{\link{upload.dataset}}; \code{\link{workspace}}
}

