% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discover.R
\name{discoverSchema}
\alias{discoverSchema}
\title{Discover web service schema.}
\usage{
discoverSchema(helpURL, scheme = "https",
  host = "ussouthcentral.services.azureml.net", api_version = "2.0")
}
\arguments{
\item{helpURL}{URL of the help page of the web service}

\item{scheme}{the URI scheme}

\item{host}{optional parameter that defaults to ussouthcentral.services.azureml.net}

\item{api_version}{AzureML API version}
}
\value{
List containing the request URL of the webservice, column names of the data, sample input as well as the input schema
}
\description{
Discover the expected input to a web service specified by a web service ID ng the workspace ID and web service ID, information specific to the consumption functions
}
\seealso{
\code{\link{publishWebService}} \code{\link{consume}} \code{\link{workspace}} \code{link{services}} \code{\link{endpoints}} \code{\link{endpointHelp}}

Other discovery functions: \code{\link{endpointHelp}},
  \code{\link{endpoints}}, \code{\link{services}},
  \code{\link{workspace}}
}
