% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/protocol.R
\name{download.datasets}
\alias{download.datasets}
\title{download.datasets Donwload one or more datasets from an AzureML workspace}
\usage{
download.datasets(datasets, ...)
}
\arguments{
\item{datasets}{One or more rows from a \code{datasets} data.frame in a workspace.}

\item{...}{Optional arguments to pass to \code{read.table} for CSV or TSV DataTypeIds. For example,
specify \code{stringsAsFactors=TRUE} if you wish, or any other valid argument to \code{read.table}.}
}
\value{
If one dataset is specified (that is, one row from a workspace \code{datasets} data.frame),
then a single data.frame is returned.
If more than one dataset is specified (more than one row), then a list of data.frames is returned.
}
\description{
Download one or more datasets from an AzureML workspace into local R
data.frame objects.
}
\note{
TSV- and CSV-formatted datasets return data.frame results with \code{stringsAsFactors=FALSE}
by default (independently of the global \code{stringsAsFactors} option).

This function can download datasets with various CSV and TSV "DataTypeIds", or "DataTypeId"
of "ARFF" or "PlainText". Other "DataTypeIds" return an error. See the AzureML Data Format
Conversion modules to convert data to a supported format.
}
\seealso{
\code{\link{workspace}}, \code{\link{datasets}}, \code{\link{read.table}},
\code{\link{download.intermediate.dataset}}
}

