% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/discover.R
\name{endpointHelp}
\alias{endpointHelp}
\title{Display AzureML Web Service Endpoint Help Screens}
\usage{
endpointHelp(e, type = c("r-snippet", "score", "jobs", "update"))
}
\arguments{
\item{e}{an AzureML web service endpoint from the \code{\link{endpoints}} function.}

\item{type}{the type of help to display.}
}
\value{
The help text is invisibly returned.
}
\description{
Download and display help for the specified AzureML web service endpoint.
The result is displayed as text using the pager.
}
\examples{
\dontrun{
workspace_id <- ""          # Your AzureML workspace id
authorization_token <- ""   # Your AsureML authorization token

ws <- workspace(
  id = workspace_id,
  auth = authorization_token
)

s <- services(ws)
e <- endpoints(ws, s[1,])
endpointHelp(e[1,])

}
}
\seealso{
Other discovery functions: \code{\link{discoverSchema}};
  \code{\link{endpoints}}, \code{\link{getEndpoints}};
  \code{\link{getWebServices}}, \code{\link{services}};
  \code{\link{workspace}}
}

