% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/publish.R
\name{expToEvaluate}
\alias{expToEvaluate}
\title{Create a string to be evaluated as an R expression}
\usage{
expToEvaluate(url)
}
\arguments{
\item{url}{the US South Central service management path (with the 'https://' omitted)}
}
\value{
string to be parsed and evaluated as an R expression.
}
\description{
Helper function that returns a string to be parsed as an R expression and evaluated inside a function. The R expression checks if the user has omitted required arguments and replaces them with the workspaces settings, if such settings were defined.
}
\keyword{internal}

