% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{get_dataset}
\alias{get_dataset}
\title{Internal function that retrieves a dataset from AzureML.}
\usage{
get_dataset(x, h, quote = "\\"", ...)
}
\arguments{
\item{x}{a list or data.frame with \code{DownloadLocation} and \code{DataTypeId} fields}

\item{h}{optional curl handle}

\item{quote}{passed to \code{\link[utils]{read.table}}}

\item{...}{additional parameters to pass to \code{read.table}}
}
\value{
a data.frame
}
\description{
Internal function that retrieves a dataset from AzureML.
}
\keyword{Internal}

