% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/internal.R
\name{get_datasets}
\alias{get_datasets}
\title{get_datasets internal function that retrieves datasets}
\usage{
get_datasets(w)
}
\arguments{
\item{w}{A workspace object}
}
\value{
a data.frame
}
\description{
internal function that retrieves datasets
}

