% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{get_default_datastore}
\alias{get_default_datastore}
\title{Get default datastore associated with a workspace}
\usage{
get_default_datastore(ws)
}
\arguments{
\item{ws}{workspace object}
}
\description{
Get default datastore associated with a workspace
}
