% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/internal.R
\name{get_experiments}
\alias{get_experiments}
\title{get_experiments internal function that retrieves experiments}
\usage{
get_experiments(w)
}
\arguments{
\item{w}{A workspace object}
}
\value{
a data.frame
}
\description{
internal function that retrieves experiments
}

