% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/discover.R
\name{getDetailsFromUrl}
\alias{getDetailsFromUrl}
\title{Helper function to extract information from a help page URL}
\usage{
getDetailsFromUrl(url)
}
\arguments{
\item{url}{the URL of a help page}
}
\value{
a vector containing the workspace ID, webservices ID and endpoint ID
}
\description{
Given a Microsoft Azure Machine Learning web service endpoint, extracts the endpoint ID and the workspace ID
}
\keyword{internal}

