% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/discover.R
\name{getEPDetails}
\alias{getEPDetails}
\title{Get Endpoint Details.}
\usage{
getEPDetails(wkID, authToken, wsID, epName, url = prodURL)
}
\arguments{
\item{wkID}{workspace ID}

\item{authToken}{primary authorization token}

\item{wsID}{the web service ID}

\item{epName}{endpoint name}

\item{url}{the API url to make the call to, by default hits the Azure management API}
}
\value{
Returns a list with the following fields:

\itemize{
 \item Name
 \item Description
 \item CreationTime
 \item WorkspaceId
 \item WebServiceId
 \item HelpLocation
 \item PrimaryKey
 \item SecondaryKey
 \item ApiLocation
 \item Version
 \item MaxConcurrentCalls
 \item DiagnosticsTraceLevel
 \item ThrottleLevel
 }
}
\description{
Get detailed information about a specific endpoint for a web service specified by the web service ID
}
\seealso{
Other discovery functions: \code{\link{discoverSchema}};
  \code{\link{getEndpoints}}; \code{\link{getFramework}};
  \code{\link{getWSDetails}}; \code{\link{getWebServices}}
}

