% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/discover.R
\name{getFramework}
\alias{getFramework}
\title{Abstraction for making GET requests.}
\usage{
getFramework(tUrl, authToken, verbose = TRUE, ...)
}
\arguments{
\item{tUrl}{the API URL}

\item{authToken}{the authentication token}

\item{verbose}{verbosity of RCurl::curlPerform}

\item{...}{additional arguments passed to RCurl::curlPerform}
}
\value{
the response as a named list
}
\description{
Framework for making GET requests to the Azure management APIs.
}
\seealso{
Other discovery functions: \code{\link{discoverSchema}};
  \code{\link{getEPDetails}}; \code{\link{getEndpoints}};
  \code{\link{getWSDetails}}; \code{\link{getWebServices}}
}
\keyword{internal}

