% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/publish.R
\name{getFunctionString}
\alias{getFunctionString}
\title{Get function source code}
\usage{
getFunctionString(x)
}
\arguments{
\item{x}{name of the function to convert to a string}
}
\value{
source code of the function as a string
}
\description{
Returns the source code of a function as a string
}
\keyword{internal}

