% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods.R
\name{is.Dataset}
\alias{is.Dataset}
\title{Test if an object is an Azure ML Dataset.}
\usage{
is.Dataset(x)
}
\arguments{
\item{x}{an R object}
}
\value{
logical value, TRUE if \code{x} represents an Azure ML Dataset.
}
\description{
Test if an object is an Azure ML Dataset.
}

