% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{is.Endpoint}
\alias{is.Endpoint}
\title{Test if an object is an Azure ML Endpoint.}
\usage{
is.Endpoint(x)
}
\arguments{
\item{x}{an R object}
}
\value{
logical value, TRUE if \code{x} represents an Azure ML web service endpoint
}
\description{
Test if an object is an Azure ML Endpoint.
}

