% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{is.Service}
\alias{is.Service}
\title{Test if an object is an Azure ML Service.}
\usage{
is.Service(x)
}
\arguments{
\item{x}{an R object}
}
\value{
logical value, TRUE if \code{x} represents an Azure ML web service
}
\description{
Test if an object is an Azure ML Service.
}
