% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{is.Workspace}
\alias{is.Workspace}
\title{Test if an object is an Azure ML Workspace.}
\usage{
is.Workspace(x)
}
\arguments{
\item{x}{an R object}
}
\value{
logical value, TRUE if \code{x} represents an Azure ML workspace.
}
\description{
Test if an object is an Azure ML Workspace.
}

