% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{list_workspaces}
\alias{list_workspaces}
\title{List all workspaces that the user has access to in the specified subscription_id
parameter.The list of workspaces can be filtered based on the resource group.}
\usage{
list_workspaces(subscription_id, resource_group = NULL)
}
\arguments{
\item{subscription_id}{To list workspaces in the specified subscription ID.}

\item{resource_group}{To list workspaces in the specified resource group.
If NULL the method will list all the workspaces within the specified subscription.}
}
\description{
List all workspaces that the user has access to in the specified subscription_id
parameter.The list of workspaces can be filtered based on the resource group.
}
