% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/publish.R
\name{loadSettings}
\alias{loadSettings}
\title{Load workspace settings from ~/azureml/settings.ini.}
\usage{
loadSettings()
}
\description{
Helper function that sets the values of the workspace id, authorization token and region. The values are stored internally and are used in any function call where the user omitted one (or more) of the 3 parameters.
}
\details{
This function allows the user to read the workspace settings from a configuration file 'settings.ini' placed in a folder 'azureml' in the working directory. The file must contain the following fields:
\itemize{
\item{wsId:}{ the workspace ID obtained from Azure ML Studio}
\item{authToken:}{ primary authorization token}
\item{region:}{ region where the workspace was created. 'ussouthcentral' for South Central US, 'europewest' for West Europe, and 'asiasoutheast' for Southeast Asia}
}
followed by the "=" token and their value.
}
\examples{
\dontrun{
#In file settings.ini
[workspace]
wsId=abcdefghijklmnopqrstuvwxyz1234567
authToken=abcdefghijklmnopqrstuvwxyz1234567
region=ussouthcentral
}
}
\keyword{internal}

