% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/publish.R
\name{packDependencies}
\alias{packDependencies}
\title{Package a function's dependencies into a base64 encoded string}
\usage{
packDependencies(functionName)
}
\arguments{
\item{functionName}{function to package dependencies from}
}
\value{
list containing the guid for the rdta file and the encoded zip
}
\description{
Find a function's in-memory and package dependencies, and turn them into a base-64 encoded zip file. This string is used in the publish API call to upload dependencies to the server.
}
\keyword{internal}

