% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/publish.R
\name{publishPreprocess}
\alias{publishPreprocess}
\title{Convert input schema to API expected format.}
\usage{
publishPreprocess(argList)
}
\arguments{
\item{argList}{list of expected input parameters in the format expected by \code{\link{publishWebService}}}
}
\value{
list of the format expected by the API
}
\description{
Helper function to convert the user-friendly input and output schema parameters to the publishWebService() function to the format expected by the API.
}
\keyword{internal}

