% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/discover.R
\name{services}
\alias{getWebServices}
\alias{services}
\title{List Available Web Services.}
\usage{
services(ws, service_id, uri = "https://management-tm.azureml.net")

getWebServices(ws, service_id, uri = "https://management-tm.azureml.net")
}
\arguments{
\item{service_id}{An optional web service id. If supplied, return the web service information for just
that service id. Leave undefined to return a data.frame of all services.}

\item{uri}{The AzureML web services URI}
}
\value{
Returns a data.frame with variables:
\itemize{
  \item Id
  \item Name
  \item Description
  \item CreationTime
  \item WorkspaceId
  \item DefaultEndpointName
}
Each row of the returned data.frame corresponds to a service.
}
\description{
Return a list of web services available to the specified Microsoft Azure Machine Learning workspace.
}
\note{
\code{getWebServices} is an alias for \code{services}.
}
\examples{
\dontrun{
workspace_id <- ""          # Your AzureML workspace id
authorization_token <- ""   # Your AsureML authorization token

ws <- workspace(
  id = workspace_id,
  auth = authorization_token
)

# Equivalent:
services(ws)
getWebServices(ws)
}
}
\seealso{
Other discovery functions: \code{\link{discoverSchema}};
  \code{\link{endpoints}}, \code{\link{getEndpoints}};
  \code{\link{workspace}}
}

