% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generic.R
\name{[[.Datasets}
\alias{[[.Datasets}
\title{Access experiment or dataset by index or name/id}
\usage{
\method{[[}{Datasets}(x, i)
}
\arguments{
\item{x}{The experiments or datasets slots of a Workspace object}

\item{i}{The index of the dataset or experiment}
}
\value{
A dataset or experiment object
}
\description{
Access individual experiments or datasets in a workspace by index or by name for datasets, id for experiments
}
\details{
Both names and ids can be obtained with the function names applied to the experiments or datasets, which are elements of the a workspace object.
}
\examples{
\dontrun{
    # provide id and auth token from ML studio -> settings
    ws = workspace(id, authorization.token)
    experiments(ws)[[1]]
    datasets(ws)[[1]]
    # name from ML studio ->datasets -> my datasets or samples -> name or
    # names(ws$datsets)
    datasets(ws)[[name]]
    # id from ML studio -> experiments -> my datasets or samples -> name
    experiments(ws)[[id]]
  }
}
\seealso{
Other dataset functions: \code{\link{as.data.frame.Dataset}};
  \code{\link{as.list.Datasets}}; \code{\link{datasets}};
  \code{\link{names.Datasets}}
}

