% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish.R
\name{test_wrapper}
\alias{test_wrapper}
\title{Test the AzureML wrapper locally}
\usage{
test_wrapper(inputDF, wrapper, fun, output_names, data.frame)
}
\arguments{
\item{inputDF}{data frame}

\item{wrapper}{the AzureML R wrapper code}

\item{fun}{a function to test}

\item{output_names}{character vector of function output names}

\item{data.frame}{i/o format}
}
\description{
Test the AzureML wrapper locally
}

