% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generic.R
\name{update.Dataset}
\alias{update.Dataset}
\title{Update and existing remote data set}
\usage{
\method{update}{Dataset}(x, data, data.type = NULL, name = NULL,
  description = NULL)
}
\arguments{
\item{x}{A remote dataset object as returned by \code{create} or extracted
from a workspace experiments field with the bracket operator/}

\item{data}{A data frame or serialized data}

\item{data.type}{The format of \code{data} or to be used to serialize it.}

\item{name}{A unique name for the dataset}

\item{description}{An extended description for the dataset}
}
\description{
Arguments set to NULL are construed as "leave unchanged".
}
\examples{
\dontrun{
ws = workspace(workspace, authorization.token)
update(ws$user.datasets[[1]], iris3)
}
}

