% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generic.R
\name{as.list.Experiments}
\alias{as.list.Experiments}
\title{Convert collection of Experiments to lists}
\usage{
\method{as.list}{Experiments}(x, ...)
}
\arguments{
\item{x}{A \code{\link{experiments}} object to convert to a list}

\item{...}{Unused, just for compatibility with generic}
}
\value{
A Dataset or Experiment object
}
\description{
Convert collection of Experiments to lists
}
\examples{
\dontrun{
ws = workspace(id, authorization.token)
as.list(experiments(ws))
as.list(datasets(ws))
}
}
\seealso{
Other experiment functions: \code{\link{as.list.Datasets}};
  \code{\link{dataset}}; \code{\link{experiments}}
}

