% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generic.R
\name{dataset}
\alias{dataset}
\title{Extract datasets from experiments}
\usage{
dataset(x, node, port, data.type)
}
\arguments{
\item{x}{Experiment from which to extract a dataset}

\item{node}{Node id}

\item{port}{Port name}

\item{data.type}{Format of intermediate data set}
}
\value{
A remote dataset
}
\description{
Extract datasets from experiments
}
\seealso{
Other experiment functions: \code{\link{as.list.Datasets}},
  \code{\link{datasets}};
  \code{\link{as.list.Experiments}};
  \code{\link{experiments}}
}

