% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/protocol.R
\name{datasets}
\alias{datasets}
\title{datasets Return datasets in an AzureML workspace}
\usage{
datasets(w, filter = c("all", "my datasets", "samples"))
}
\arguments{
\item{w}{An AzureML workspace reference returned by \code{\link{workspace}}.}

\item{filter}{Optionally filter result, returing all, mine, or sample datasets.}
}
\value{
A data.frame with class \code{Datasets} listing available datasets in the workspace.
}
\description{
Return datasets in an AzureML workspace, optionally filtering on sample or my
datasets.
}
\note{
\code{datasets(w)} is equivalent to \code{w$datasets}. Since \code{w$datasets} is simply
an R data.frame, you can alternatively filter on any variable as desired.
}
\seealso{
\code{\link{workspace}}, \code{\link{experiments}}
}

