% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generic.R
\name{datasets}
\alias{datasets}
\title{Extract datasets from experiments}
\usage{
datasets(workspace)
}
\arguments{
\item{workspace}{workspace object, as returned by \code{\link{workspace}}}
}
\value{
A remote dataset
}
\description{
Extract datasets from experiments
}
\seealso{
Other dataset functions: \code{\link{[[.Datasets}};
  \code{\link{as.data.frame.Dataset}};
  \code{\link{as.list.Datasets}};
  \code{\link{names.Datasets}}
}

