% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/delete_dataset.R
\name{delete.dataset}
\alias{delete.dataset}
\title{Delete one or more datasets from an AzureML workspace.}
\usage{
delete.dataset(ws, name, ...)
}
\description{
Delete one or more datasets from an AzureML workspace.
}

