% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\name{download.intermediate.dataset}
\alias{download.intermediate.dataset}
\title{Download a dataset from an AzureML experiment port.}
\usage{
download.intermediate.dataset(ws, experiment, node_id,
  port_name = "Results dataset", data_type_id = "GenericCSV", ...)
}
\arguments{
\item{ws}{An AzureML workspace reference returned by \code{\link{workspace}}.}

\item{experiment}{AzureML experiment ID obtained from "Generate Data Access Code"}

\item{node_id}{Experiment node ID obtained from "Generate Data Access Code"}

\item{port_name}{Experiment port name obtained from "Generate Data Access Code". The default is "Results dataset"}

\item{data_type_id}{Experiment data type id obtained from "Generate Data Access Code". The default is "GenericCSV"}

\item{...}{Optional arguments to pass to \code{read.table} for CSV or TSV DataTypeIds. For example, specify \code{stringsAsFactors=TRUE} if you wish, or any other valid argument to \code{read.table}.}
}
\value{
A data frame.
}
\description{
Download a dataset from an AzureML experiment port.
}
\note{
TSV- and CSV-formatted datasets return data frame results with \code{stringsAsFactors=FALSE} by default (independently of the global \code{stringsAsFactors} option).
This function can download datasets with various CSV and TSV "DataTypeIds", or "DataTypeId" of "ARFF" or "PlainText". Other "DataTypeIds" return an error. See the AzureML Data Format Conversion modules to convert data to a supported format.
}
\seealso{
\code{\link{workspace}}, \code{\link{datasets}}, \code{\link[utils]{read.table}}
\code{\link{download.datasets}}

Other dataset functions: \code{\link{datasets}};
  \code{\link{delete.datasets}};
  \code{\link{upload.dataset}}; \code{\link{workspace}}

Other experiment functions: \code{\link{experiments}};
  \code{\link{workspace}}
}

