% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generic.R
\name{experiments}
\alias{experiments}
\title{Extract experiments from Azure ML studio}
\usage{
experiments(workspace)
}
\arguments{
\item{workspace}{workspace object, as returned by \code{\link{workspace}}}
}
\description{
Extract experiments from Azure ML studio
}
\seealso{
Other experiment functions: \code{\link{as.list.Datasets}};
  \code{\link{as.list.Experiments}}; \code{\link{dataset}}
}

