% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/protocol.R
\name{experiments}
\alias{experiments}
\title{experiments Return experiments in an AzureML workspace}
\usage{
experiments(w, filter = c("all", "my datasets", "samples"))
}
\arguments{
\item{w}{An AzureML workspace reference returned by \code{\link{workspace}}.}

\item{filter}{Optionally filter result, returing all, mine, or sample experiments.}
}
\value{
A data.frame with class \code{Experiments} listing available experiments in the workspace.
}
\description{
Return experiments in an AzureML workspace, optionally filtering on sample or my
experiments.
}
\note{
\code{experiments(w)} is equivalent to \code{w$experiments}. Since \code{w$experiments} is
simply an R data.frame, you can alternatively filter on any variable as desired.
}
\seealso{
\code{\link{workspace}}, \code{\link{datasets}}
}

