% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/internal.R
\name{get_datasets}
\alias{get_datasets}
\title{Internal function that retrieves datasets.}
\usage{
get_datasets(ws)
}
\arguments{
\item{ws}{A workspace object}
}
\value{
a data.frame
}
\description{
Internal function that retrieves datasets.
}
\keyword{Internal}

