% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{get_experiments}
\alias{get_experiments}
\title{Internal function that retrieves experiments.}
\usage{
get_experiments(ws)
}
\arguments{
\item{ws}{A workspace object}
}
\value{
a data.frame
}
\description{
Internal function that retrieves experiments.
}
\keyword{Internal}

