% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/consume.R
\name{getDetailsFromUrl}
\alias{getDetailsFromUrl}
\title{Helper function to extract information from a help page URL}
\usage{
getDetailsFromUrl(helpURL)
}
\arguments{
\item{helpURL}{the URL of a help page}
}
\value{
a list containing the endpoint ID and the workspace ID
}
\description{
Given a Microsoft Azure Machine Learning web service endpoint, extracts the endpoint ID and the workspace ID
}
\keyword{internal}

