% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/internal.R
\name{try_fetch}
\alias{try_fetch}
\title{Try to fetch a uri/handle, retrying on certain returned status codes after a timeout}
\usage{
try_fetch(uri, handle, retry_on = c(503, 504, 509, 400, 401, 440),
  tries = 3, delay = 10, exponent = 1.2)
}
\arguments{
\item{uri}{the uri to fetch}

\item{handle}{a curl handle}

\item{retry_on}{HTTP status codes that result in retry}

\item{tries}{number of tries before failing}

\item{delay}{in seconds between retries, subject to exponent}

\item{exponent}{increment each successive delay by delay^exponent}
}
\value{
the result of curl_fetch_memory(uri, handle)
}
\description{
Try to fetch a uri/handle, retrying on certain returned status codes after a timeout
}

