% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/protocol.R
\name{upload.dataset}
\alias{upload.dataset}
\title{upload.dataset Upload an R data.frame to an AzureML workspace}
\usage{
upload.dataset(x, w, name, description = "", family_id = "", ...)
}
\arguments{
\item{x}{An R data.frame object}

\item{w}{A reference to an AzureML workspace returned by \code{workspace}}

\item{name}{A character name for the new AzureML dataset (may not match an existing dataset name)}

\item{description}{An optional character description of the dataset}

\item{family_id}{An optional AzureML family identifier}

\item{...}{Optional additional options passed to \code{write.table}}
}
\value{
A single-row data.frame of "Datasets" class that corresponds to the uploaded
  object now available in w$datasets.
}
\description{
Upload any R data.frame to an AzureML workspace using the "GenericTSV" format.
}
\note{
The additional \code{write.table} options may not include \code{sep} or \code{row.names} or \code{file},
      but any other options are accepted.
The AzureML API does not support uploads for _replacing_ datasets with new data by re-using a name. If
you need to do this, first delete the dataset from the AzureML Studio interface, then upload a new version.
}

