% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/protocol.R
\name{workspace}
\alias{workspace}
\title{workspace Create a reference to an AzureML Studio workspace}
\usage{
workspace(id, auth, api_endpoint = "https://studio.azureml.net",
  management_endpoint = "https://management.azureml.net",
  config = "~/.azureml/settings.json")
}
\arguments{
\item{id}{Optional workspace id from ML studio -> settings -> WORKSPACE ID}

\item{auth}{Optional authorization token from ML studio -> settings -> AUTHORIZATION TOKENS}

\item{api_endpoint}{Optional AzureML API web service URI}

\item{management_endpoint}{Optional AzureML management web service URI}

\item{config}{Optional settings file containing id and authorization info. Only used if \code{id} and \code{auth} are missing.}
}
\value{
An R environment of class "Workspace" containing at least the following objects:
\describe{
  \item{experiments}{Collection of experiments in the workspace represented as an "Experiments" data.frame}
  \item{datasets}{Collection of datasets in the workspace represented as a "Datasets" data.frame}
}
}
\description{
Create a reference to an AzureML Studio workspace. Data corresponding
to experiments and datasets in the workspace are cached in the result.
See \code{\link{refresh}} to update cached data.
}
\note{
If the \code{id} and \code{auth} parameters are missing, the function
attempts to read values from the \code{config} file, JSON formatted as
shown in \url{https://github.com/RevolutionAnalytics/azureml/issues/13}.
}
\seealso{
\code{\link{datasets}}, \code{\link{experiments}}, \code{\link{refresh}}
}

