% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/workspace.R
\name{workspace}
\alias{workspace}
\title{Create a reference to an AzureML Studio workspace.}
\usage{
workspace(id, auth, api_endpoint, management_endpoint,
  config = getOption("AzureML.config"), ..., .validate = TRUE)
}
\arguments{
\item{id}{Optional workspace id from ML studio -> settings -> WORKSPACE ID}

\item{auth}{Optional authorization token from ML studio -> settings -> AUTHORIZATION TOKENS}

\item{api_endpoint}{Optional AzureML API web service URI. Defaults to \url{https://studio.azureml.net} if not provided and not specified in config.  See note.}

\item{management_endpoint}{Optional AzureML management web service URI. Defaults to \url{https://management.azureml.net} if not provided and not specified in config.  See note.}

\item{config}{Optional settings file containing id and authorization info. Used if any of the other arguments are missing. The default config file is \code{~/.azureml/settings.json}, but you can change this location by setting \code{options(AzureML.config = "newlocation")}}

\item{...}{ignored}

\item{.validate}{If TRUE}
}
\value{
An R environment of class \code{Workspace} containing at least the following objects:
\describe{
  \item{experiments}{Collection of experiments in the workspace represented as an \code{Experiments} data.frame}
  \item{datasets}{Collection of datasets in the workspace represented as a \code{Datasets} data.frame}
  \item{services}{Collection of web services in the workspace represented as a data.frame}
}
}
\description{
Create a reference to an AzureML Studio workspace, returning a \code{Workspace} object that is an R environment containing details and data associated with the AzureML work space. Data corresponding to services, experiments, and datasets in the workspace are cached in the result object environment. See \code{\link{refresh}} about updating cached data.
}
\note{
If any of the \code{id}, \code{auth}, \code{api_endpoint} or \code{management_endpoint} arguments are missing, the function attempts to read values from the \code{config} file with JSON format: \preformatted{
 {"workspace":{
   "id": "test_id",
   "authorization_token": "test_token",
   "api_endpoint": "api_endpoint",
   "management_endpoint": "management_endpoint"
 }}
} (see \url{https://github.com/RevolutionAnalytics/azureml/issues/13}).
}
\seealso{
\code{\link{datasets}}, \code{\link{experiments}}, \code{\link{refresh}},
         \code{\link{services}}, \code{\link{consume}}, \code{\link{publishWebService}}

Other consumption functions: \code{\link{consume}}

Other dataset functions: \code{\link{datasets}};
  \code{\link{delete.datasets}};
  \code{\link{download.intermediate.dataset}};
  \code{\link{upload.dataset}}

Other discovery functions: \code{\link{discoverSchema}};
  \code{\link{endpointHelp}}; \code{\link{endpoints}},
  \code{\link{getEndpoints}}; \code{\link{getWebServices}},
  \code{\link{services}}

Other experiment functions: \code{\link{download.intermediate.dataset}};
  \code{\link{experiments}}

Other publishing functions: \code{\link{deleteWebService}};
  \code{\link{publishWebService}},
  \code{\link{updateWebService}}
}

